fix(rules): match spec for defaulted required args
This adds a new test, that fails on master, alongside with the correct changes
for it to work as expected.

The validation was failing for arguments that are defined as required, but have
a default value, preventing a non-required argument to override their value,
while it should work as expected.